clean up/relocate convex utils
