+++ convex/_generated/api.d.ts
@@ -35,6 +35,7 @@ import type * as lib_clerk from "../lib/clerk.js";
 import type * as lib_fetch from "../lib/fetch.js";
 import type * as lib_modelTags from "../lib/modelTags.js";
 import type * as lib_sharp from "../lib/sharp.js";
+import type * as lib_utils from "../lib/utils.js";
 import type * as lib_valibot from "../lib/valibot.js";
 import type * as migrations from "../migrations.js";
 import type * as rules from "../rules.js";
@@ -43,7 +44,6 @@ import type * as shared_helpers from "../shared/helpers.js";
 import type * as shared_utils from "../shared/utils.js";
 import type * as types from "../types.js";
 import type * as users from "../users.js";
-import type * as utils from "../utils.js";
 import type * as workflows_actions_evaluateNsfwProbability from "../workflows/actions/evaluateNsfwProbability.js";
 import type * as workflows_actions_generateImageCaption from "../workflows/actions/generateImageCaption.js";
 import type * as workflows_actions_textToImage_fal from "../workflows/actions/textToImage/fal.js";
@@ -89,6 +89,7 @@ declare const fullApi: ApiFromModules<{
   "lib/fetch": typeof lib_fetch;
   "lib/modelTags": typeof lib_modelTags;
   "lib/sharp": typeof lib_sharp;
+  "lib/utils": typeof lib_utils;
   "lib/valibot": typeof lib_valibot;
   migrations: typeof migrations;
   rules: typeof rules;
@@ -97,7 +98,6 @@ declare const fullApi: ApiFromModules<{
   "shared/utils": typeof shared_utils;
   types: typeof types;
   users: typeof users;
-  utils: typeof utils;
   "workflows/actions/evaluateNsfwProbability": typeof workflows_actions_evaluateNsfwProbability;
   "workflows/actions/generateImageCaption": typeof workflows_actions_generateImageCaption;
   "workflows/actions/textToImage/fal": typeof workflows_actions_textToImage_fal;
